add stf3 test

+++ t2i23-test-schnell.py
@@ -15,7 +15,9 @@ torch.cuda.empty_cache()
 
 # t2i_pipe = DiffusionPipeline.from_pretrained("stable-diffusion-v1-5/stable-diffusion-v1-5").to("cuda")
 t2i_pipe = DiffusionPipeline.from_pretrained("Keffisor21/flux1-schnell-bnb-nf4").to("cuda")
-
+t2i_pipe.transformer = t2i_pipe.transformer.half()
+t2i_pipe.vae = t2i_pipe.vae.half()
+t2i_pipe.text_encoder = t2i_pipe.text_encoder.half()
 pipeline = TrellisImageTo3DPipeline.from_pretrained("microsoft/TRELLIS-image-large")
 pipeline.cuda()
